requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
python-dotenv==1.0.0

# AI/ML dependencies
torch>=2.0.0
transformers==4.35.0
diffusers==0.23.0
accelerate==0.24.0

# 3D processing
trimesh==4.0.5
numpy-stl==3.0.1
# open3d>=0.16.0  # TODO: Add when Python 3.13 support available
# pymeshlab>=2022.2  # TODO: Add when Python 3.13 support available
scikit-image>=0.21.0

# Utilities
numpy==1.24.3
pillow==10.1.0
requests==2.31.0
pydantic==2.5.0
aiofiles==23.2.1

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0